python_backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
google-cloud-vision==3.4.3
xero-python==1.19.0
scikit-learn==1.3.0
pandas==2.1.0
celery==5.3.1
aiosmtplib==2.0.2
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1 